challenges/5_Van_Halen_Radiation_Belt/1_Meet_Me_At_Midway/solver/requirements.txt
numpy
pwntools